Refactor block structure
